I added the line skipping function
My logic behind this function was that the code would end and stop running once it found a non valid expr. Instead what I did was, I replaced the default case in the lookup functions token with '\n' instead of EOF. Then in the lex function, if the char class was unknown, I would use a helper function i created that would call getChar until the next char variable was equal to the new line character. once it found the new line character, the program would then outpue that it was an invalid expression and move onto the next expression.